# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "contact": {}
  },
  "paths": {
...
    "/api/v1/sync": {
      "post": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Enqueues a background sync job for the authenticated user",
        "tags": [
          "sync"
        ],
        "summary": "Trigger ad-hoc sync",
        "responses": {
          "202": {
            "description": "Accepted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/internal_api.ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/internal_api.ErrorResponse"
                }
              }
            }
          }
        },
        "parameters": []
      }
    },
    "/api/v1/sync/status": {
      "get": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Consultar o estado persistido da sincronização",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/internal_api.SyncStatusResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/internal_api.ErrorResponse"
                }
              }
            }
          }
        },
        "parameters": []
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "github_com_arvind_whoop-stats_internal_db.SyncStatus": {
        "type": "object",
        "properties": {
          "error_message": {
            "type": "string",
            "nullable": true
          },
          "finished_at": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "last_success_at": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "resource": {
            "type": "string"
          },
          "started_at": {
            "type": "string",
            "format": "date-time"
          },
          "state": {
            "type": "string"
          },
          "user_id": {
            "type": "string"
          }
        },
        "required": [
          "error_message",
          "finished_at",
          "last_success_at",
          "resource",
          "started_at",
          "state",
          "user_id"
        ]
      },
...
      "internal_api.ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              }
            }
          }
        },
        "required": [
          "error"
        ]
      },
...
      "internal_api.SyncStatusResponse": {
        "type": "object",
        "properties": {
          "resources": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/github_com_arvind_whoop-stats_internal_db.SyncStatus"
            }
          },
          "running": {
            "type": "boolean"
          }
        },
        "required": [
          "resources",
          "running"
        ]
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}